+++ flask_security/core.py
@@ -137,7 +137,7 @@ _default_config: dict[str, t.Any] = {
     "SUBDOMAIN": None,
     "FLASH_MESSAGES": True,
     "RETURN_GENERIC_RESPONSES": False,
-    "USE_REGISTER_V2": False,
+    "USE_REGISTER_V2": True,
     "I18N_DOMAIN": "flask_security",
     "I18N_DIRNAME": "builtin",
     "EMAIL_VALIDATOR_ARGS": None,
@@ -1239,6 +1239,10 @@ class Security:
     .. versionadded:: 5.6.0
         ``username_recovery_form``, ``change_username_form``,
 
+    .. versionchanged:: 5.7.0
+        ``register_form`` default value is RegisterFormV2 and
+        ``confirm_register_form`` default is now ``None``.
+
     .. deprecated:: 4.0.0
         ``send_mail`` and ``send_mail_task``. Replaced with ``mail_util_cls``.
         ``two_factor_verify_password_form`` removed.
@@ -1263,8 +1267,8 @@ class Security:
         verify_form: t.Type[VerifyForm] = VerifyForm,
         change_email_form: t.Type[ChangeEmailForm] = ChangeEmailForm,
         change_username_form: t.Type[ChangeUsernameForm] = ChangeUsernameForm,
-        confirm_register_form: t.Type[ConfirmRegisterForm] = ConfirmRegisterForm,
-        register_form: t.Type[RegisterForm] | t.Type[RegisterFormV2] = RegisterForm,
+        confirm_register_form: t.Type[ConfirmRegisterForm] | None = None,
+        register_form: t.Type[RegisterForm] | t.Type[RegisterFormV2] = RegisterFormV2,
         forgot_password_form: t.Type[ForgotPasswordForm] = ForgotPasswordForm,
         reset_password_form: t.Type[ResetPasswordForm] = ResetPasswordForm,
         change_password_form: t.Type[ChangePasswordForm] = ChangePasswordForm,
@@ -1511,9 +1515,8 @@ class Security:
             ):
                 self.forms[form_name].cls = form_cls
 
-        self._use_confirm_form = True
         # deprecate confirm_register_form, ConfirmRegisterForm and RegisterForm
-        if self.forms["confirm_register_form"].cls != ConfirmRegisterForm:
+        if self.forms["confirm_register_form"].cls:
             warnings.warn(
                 "The ConfirmRegisterForm and the confirm_register_form"
                 " option are"
@@ -1522,10 +1525,8 @@ class Security:
                 DeprecationWarning,
                 stacklevel=2,
             )
-        if (
-            self.forms["register_form"].cls
-            and self.forms["register_form"].cls != RegisterForm
-            and issubclass(self.forms["register_form"].cls, RegisterForm)
+        if self.forms["register_form"].cls and issubclass(
+            self.forms["register_form"].cls, RegisterForm
         ):
             warnings.warn(
                 "The RegisterForm is"
@@ -1534,15 +1535,20 @@ class Security:
                 DeprecationWarning,
                 stacklevel=2,
             )
-        if cv("USE_REGISTER_V2", app=app):
+        if not cv("USE_REGISTER_V2", app=app):
+            warnings.warn(
+                "The SECURITY_USE_REGISTER_V2 configuration option is"
+                " deprecated as of version 5.7.0 and will be removed in a future"
+                " release.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            # Switch back to old register and confirm_register forms
             # Only do this is they haven't subclassed
-            if self.forms["register_form"].cls == RegisterForm:
-                self.forms["register_form"].cls = RegisterFormV2
-                self.forms["confirm_register_form"].cls = None
-        if self.forms["register_form"].cls and issubclass(
-            self.forms["register_form"].cls, RegisterFormV2
-        ):
-            self._use_confirm_form = False
+            if self.forms["register_form"].cls == RegisterFormV2:
+                self.forms["register_form"].cls = RegisterForm
+            if not self.forms["confirm_register_form"].cls:
+                self.forms["confirm_register_form"].cls = ConfirmRegisterForm
 
         # The following will be set as attributes and initialized from constructor or
         # kwargs or config.