+++ flask_security/forms.py
@@ -675,6 +675,9 @@ class ConfirmRegisterForm(Form, RegisterFormMixin, UniqueEmailFormMixin):
     to still return important bad-password messages.
     In the case of an existing email or username - we set form.existing_xx so that
     the view can decide how to match responses (e.g. json responses always return 200).
+
+    .. deprecated:: 5.6.0
+        Replaced with RegisterFormV2
     """
 
     # Password optional when Unified Signin enabled.
@@ -723,6 +726,12 @@ class ConfirmRegisterForm(Form, RegisterFormMixin, UniqueEmailFormMixin):
 
 
 class RegisterForm(ConfirmRegisterForm, NextFormMixin):
+    """Register Form
+
+    .. deprecated:: 5.6.0
+        Replaced with RegisterFormV2
+    """
+
     # Password optional when Unified Signin enabled.
     password_confirm = PasswordField(
         get_form_field_label("retype_password"),