+++ flask_security/views.py
@@ -290,11 +290,9 @@ def logout():
 def register() -> ResponseValue:
     """View function which handles a registration request."""
 
-    # For some unknown historic reason - if you don't require confirmation
-    # (via email) then you need to type in your password twice. That might
-    # make sense if you can't reset your password but in modern (2020) UX models
-    # don't ask twice.
-    if (_security.confirmable or request.is_json) and _security._use_confirm_form:
+    if (_security.confirmable or request.is_json) and _security.forms[
+        "confirm_register_form"
+    ].cls:
         form_name = "confirm_register_form"
     else:
         form_name = "register_form"